project: fix code style issue's
Fix various code style and compiler warnings. The most frequently
violated code style issue is the 'no spaces after control blocks' rule
(i.e. `if()' instead of `if ()').

Various compiler warnings for the msvc debug builds have also been
fixed.

+++ source/ipv4/ip-output.c
@@ -35,7 +35,8 @@ void ipv4_output(struct netbuf *nb, uint32_t dst)
 #endif
 
 	header->tos = 0;
-	header->length = htons(nb->network.size + nb->transport.size + nb->application.size);
+	header->length = htons((uint16_t)(nb->network.size +
+		nb->transport.size + nb->application.size));
 	header->offset = 0;
 
 	if(proto == IP_PROTO_IGMP)
